исправлены замечания по заданию 4.1
